+++ lib/easytest/matcher/contain_exactly.rb
@@ -1,0 +1,13 @@
+module Easytest
+  module Matcher
+    class ContainExactly < Base
+      def match?
+        (actual.to_a - expected).empty?
+      end
+
+      def message
+        "contain exactly"
+      end
+    end
+  end
+end