+++ test/smoke_test.rb
@@ -14,6 +14,7 @@ test "simple case" do
   expect(123).to_be_instance_of Integer
   expect({ a: 1 }).to_include :a
   expect(/^fo/).to_match "foo"
+  expect(1..4).to_contain_exactly 1, 3, 4, 2
   expect { 1 / 0 }.to_raise ZeroDivisionError
   expect { 1 / 2 }.to_raise_nothing
 end